fixed a small floating type bug in the matching.py

+++ opensfm/matching.py
@@ -440,12 +440,14 @@ def _match_descriptors_impl(
             matches = match_brute_force_symmetric(d1, d2, overriden_config)
         else:
             matches = match_brute_force(d1, d2, overriden_config)
-# CUSTOM CHANGES #7 BEGIN
+    # CUSTOM CHANGES #7 BEGIN
     elif matcher_type == "LIGHTGLUE":
-        p1 = features_data1.points
-        p2 = features_data2.points
-        matches = match_lightglue(d1, d2, p1, p2, overriden_config["feature_type"].upper())
-# CUSTOM CHANGES #7 END
+        p1 = features_data1.points.astype(np.float32)
+        p2 = features_data2.points.astype(np.float32)
+        matches = match_lightglue(
+            d1, d2, p1, p2, overriden_config["feature_type"].upper()
+        )
+    # CUSTOM CHANGES #7 END
     else:
         raise ValueError("Invalid matcher_type: {}".format(matcher_type))
 
@@ -467,6 +469,7 @@ def _match_descriptors_impl(
         np.array(matches, dtype=int),
         matcher_type,
     )
+
 
 def match_robust(
     im1: str,
@@ -784,15 +787,21 @@ def match_brute_force_symmetric(
 
     return list(set(matches_ij).intersection(set(matches_ji)))
 
+
 # CUSTOM CHANGES #6 BEGIN
 
-def match_lightglue(d1: np.ndarray, d2: np.ndarray, p1: np.ndarray, p2: np.ndarray, feature_type: str) -> List[Tuple[int, int]]:
+
+def match_lightglue(
+    d1: np.ndarray, d2: np.ndarray, p1: np.ndarray, p2: np.ndarray, feature_type: str
+) -> List[Tuple[int, int]]:
     if (
         not hasattr(match_lightglue, "lightglue")
         or match_lightglue.feature_type != feature_type.lower()
     ):
         if feature_type.lower() not in ("disk", "superpoint"):
-            raise RuntimeError(f"LightGlue: Unsupported feature type {feature_type}. LightGlue only supports DISK and SUPERPOINT")
+            raise RuntimeError(
+                f"LightGlue: Unsupported feature type {feature_type}. LightGlue only supports DISK and SUPERPOINT"
+            )
         if not torch.cuda.is_available():
             logger.error("LightGlue requires CUDA")
             raise RuntimeError("LightGlue requires CUDA")
@@ -808,7 +817,7 @@ def match_lightglue(d1: np.ndarray, d2: np.ndarray, p1: np.ndarray, p2: np.ndarr
         #     match_lightglue.local.lightglue
         # )
         match_lightglue.feature_type = feature_type
-    logger.debug("LightGlue initialized")
+        logger.debug("LightGlue initialized")
 
     feats0full = {
         "keypoints": torch.tensor(p1)[None, :].cuda(),
@@ -820,9 +829,7 @@ def match_lightglue(d1: np.ndarray, d2: np.ndarray, p1: np.ndarray, p2: np.ndarr
         "descriptors": torch.tensor(d2)[None, :].cuda(),
     }
 
-    matches01 = match_lightglue.lightglue(
-        {"image0": feats0full, "image1": feats1full}
-    )
+    matches01 = match_lightglue.lightglue({"image0": feats0full, "image1": feats1full})
 
     # in opensfm matches are of type List[Tuple[int, int]], where each tuple is a pair of indices into the keypoints
     # in lightglue matches are of type torch.Tensor, where each row is a pair of indices into the keypoints
@@ -831,7 +838,6 @@ def match_lightglue(d1: np.ndarray, d2: np.ndarray, p1: np.ndarray, p2: np.ndarr
     matches = [tuple(x) for x in matches01]
 
     return matches
-
 
 
 # CUSTOM CHANGES #6 END
